Nick/animator interface (#134)
* making animations interpolate w.r.t the time of the first key frame

* property for animations being played over time or not, can set current frame of an animation

* starting to work on interface for playing animations

* merge fix

* animation works

* animation changes

* animation fix

* file browser

* animation fix initialization lol

* animation fix again

* Time fix for the FFT water

+++ Source/YTE/Graphics/FFT_WaterSimulation.hpp
@@ -207,7 +207,7 @@ namespace YTE
     void SetWaveHeight(float aWaveHeight);
     void SetWindFactor(glm::vec2 aWindFactor);
     void SetVertexDistance(glm::vec2 aDistance);
-    void SetTimeDilationEffect(float aTimeDilationEffect);
+    void SetTimeDilationEffect(double aTimeDilationEffect);
     void SetReset(bool aReset);
     void SetInstancingAmount(int aAmount);
     void SetUseFFT(bool aValue);
@@ -239,7 +239,7 @@ namespace YTE
     glm::vec2 GetVertexDistance();
     float GetVertexDistanceX();
     float GetVertexDistanceZ();
-    float GetTimeDilationEffect();
+    double GetTimeDilationEffect();
     bool GetReset();
     int GetInstancingAmount();
     bool GetUseFFT();
